moved all hybrid controller examples directly to default impedance controller examples file
